Football players market value prediction algorithm
Fifa 22 dataset is used for this study and for the popularity metrics, data is scraped from Wikipedia. The gathered data is cleaned and transformed and then the exploratory analysis is performed on the data to get insights from the dataset. Feature selection is done on the basis of the correlation matrix. The final transformed dataset is then used for model building. Models are created using linear regression, random forest, and gradient boosting algorithms. The model is evaluated in three different scenarios, first with players’ performance and demographics data, second with the addition of players’ popularity data, and third by adding
players’ salary averages. Three different models are created for each scenario. The models are optimized using cross-validation and hyperparameter tuning. The best-performing model is selected on the basis of the highest R square and least mean square value.